services/parser/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.1
hl7==0.4.5
pydicom==2.4.3
fhir.resources==7.1.0